added more ciphers
Former-commit-id: 502e1461fd98aeab4f3dd08b6937c8f77c340487

+++ ssl/cipher_suites.go
@@ -195,7 +195,7 @@ const (
 type cipherSuite struct {
 	name            string
 	opensslname     string
-	MinProtoVersion uint16
+	MinProtoVersion int
 	keyExhange      string
 	authentication  string
 	encryption      string
